move spack.util.web.get_checksums_for_versions() to spack.stage
Function moved to break another import cycle:
  - spack.main
  - spack.architecture
  - spack.compiler
  - spack.repo
  - spack.caches
  - spack.fetch_strategy
  - spack.util.web
    (This commit breaks the cycle at this point.
  - spack.stage
  - spack.mirror
  - spack.fetch_strategy (cycle!)

My reasoning for breaking the cycle in this way is
that it seems questionable from a software
architecture standpoint to have something in a
utility module (spack.util.web), which I think of
as operating at a lower abstraction layer, reach
"back out" for code in spack.stage, which I
consider to be operating at a higher abstraction
layer.

The get_checksums_for_versions() function is the
only function in spack.util.web that requires
code from spack.stage, and it only seems to use
this code as a convenient download mechanism; not
for anything specific to handling stages.

In this case, either:
  1 - the get_checksums_for_versions() function
      should be moved to a higher-level module,

  2 - the parts of spack.stage that handle
      downloads should be demoted to a utility
      module, or

  3 - some compromise between the above options,
      where only the code in spack.stage that
      handle downloads is properly refactored into
      a utility module, and both spack.stage and
      spack.util.web are rewritten to use the
      newly-created common utility.

In this commit, I went with (1), though (3) is
arguably a better way to maintain the software
architecture long-term.
